requirements.txt
pytesseract
easyocr
pdf2image
opencv-python
pillow
numpy
